pyproject.toml
[project]
name = "hollowfoot"
version = "0.0.0"
authors = [
  { name="Mark Wolfman", email="wolfman@anl.gov" },
]
description = "Companion workflow engine for xraylarch."
readme = "README.rst"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering",
]
keywords = []
dependencies = [
    "numpy",
    "xraylarch",
]

[dependency-groups]
dev = ["black", "pytest", "ruff", "isort", "mypy"]

[project.urls]
Homepage = "https://github.com/spc-group/hollowfoot"

[project.scripts]
# command = "hollowfoot.__main__:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.isort]
profile = "black"

[tool.mypy]
files = ["src/"]
follow_imports = "skip"
follow_untyped_imports = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82", "F401"]
